+++ lang/en/local_boostnavigation.php
@@ -90,6 +90,15 @@ Further information to the parameters:
 </ul>
 Please note:
 <ul>
+<li>The title parameter can contain placeholders, for example {coursefullname} to create a node labeled with the current course\'s full name. Placeholders are encapsulated in curly brackets and will be replaced automatically when the custom node is created.<br />Available placeholders are:
+<ul>
+<li>{coursefullname}: The course\'s full name</li>
+<li>{courseshortname}: The course\'s shortname</li>
+<li>{editingtoggle}: The value \'Turn editing on\' or \'Turn editing off\' from the currently used language pack</li>
+<li>{userusername}: The logged in user\'s full name</li>
+<li>{userusername}: The logged in user\'s username</li>
+</ul>
+</li>
 <li>The link parameter can contain placeholders, for example /course/edit.php?id={courseid} to create a node linking to the current course\'s settings page. Placeholders are encapsulated in curly brackets and will be replaced automatically when the custom node is created.<br />Available placeholders are:
 <ul>
 <li>{courseid}: The course\'s (internal) ID</li>
@@ -122,6 +131,14 @@ Further information to the parameters:
 Please note:
 <ul>
 <li>Pipe dividing for optional parameters is always needed if they are located between other options. This means that you have to separate params with the pipe character although they are empty. Also see the example for the Faculty of mathematics custom node above.</li>
+<li>The title parameter can contain placeholders, for example {coursefullname} to create a node labeled with the current course\'s full name. Placeholders are encapsulated in curly brackets and will be replaced automatically when the custom node is created.<br />Available placeholders are:
+<ul>
+<li>{coursefullname}: The course\'s full name</li>
+<li>{courseshortname}: The course\'s shortname</li>
+<li>{editingtoggle}: The value \'Turn editing on\' or \'Turn editing off\' from the currently used language pack</li>
+<li>{userusername}: The logged in user\'s full name</li>
+<li>{userusername}: The logged in user\'s username</li>
+</ul>
 <li>The link parameter can contain placeholders, for example /course/edit.php?id={courseid} to create a node linking to the current course\'s settings page. Placeholders are encapsulated in curly brackets and will be replaced automatically when the custom node is created.<br />Available placeholders are:
 <ul>
 <li>{courseid}: The course\'s (internal) ID</li>